Updated delete user

+++ src/app/modules/user/user.controller.ts
@@ -7,6 +7,7 @@ import pick from '../../../shared/pick';
 import sendResponse from '../../../shared/sendResponse';
 import { userFilterableFields } from './user.constant';
 import { UserService } from './user.service';
+import { Organization } from '../organization/organization.model';
 
 const addUser: RequestHandler = catchAsync(
   async (req: Request, res: Response) => {
@@ -114,6 +115,18 @@ const updateUser: RequestHandler = catchAsync(
 const deleteUser: RequestHandler = catchAsync(
   async (req: Request, res: Response) => {
     const organization_id = req.user?.organization_id;
+    const organization = await Organization.findOne({
+      _id: organization_id,
+    }).select('user_delete_permission');
+
+    if (!organization?.user_delete_permission) {
+      return sendResponse(res, {
+        statusCode: httpStatus.UNAUTHORIZED,
+        success: false,
+        message: 'You do not have permission to delete user',
+      });
+    }
+
     const id = req.params.id;
     const result = await UserService.deleteUser(id, organization_id);
     sendResponse(res, {
